requirements.txt
future==0.15.2
pycrypto==2.6.1
urllib3==1.16
